requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
beautifulsoup4==4.10.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==2.0.12
click==8.1.8
coverage==7.8.0
cryptography==45.0.2
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
et_xmlfile==2.0.0
eventlet==0.33.0
exceptiongroup==1.2.2
filelock==3.18.0
Flask==3.0.3
flask-cors==6.0.0
fonttools==4.58.0
fpdf2==2.8.3
fsspec==2025.3.2
gevent==25.5.1
greenlet==3.2.2
gunicorn==20.1.0
h11==0.16.0
html2text==2020.1.16
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
Levenshtein==0.27.1
lxml==5.1.1
MarkupSafe==3.0.2
numpy==2.2.6
openai==1.77.0
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pdfminer.six==20211012
pillow==11.2.1
pluggy==1.6.0
py==1.11.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyMuPDF==1.25.5
PyPDF2==1.26.0
pytest==6.2.5
pytest-cov==2.12.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-Levenshtein==0.27.1
python-magic==0.4.24
pytz==2021.3
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.26.0
safetensors==0.5.3
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
thefuzz==0.22.1
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tqdm==4.62.3
transformers==4.52.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==1.26.7
validators==0.18.2
Werkzeug==3.1.3
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
